Change type of default twitter time check for new tweets

+++ src/Lykke.blue.Service.InspireStream.Core/AppSettings.cs
@@ -20,7 +20,7 @@
 
     public class TwitterSettings
     {
-        public int DefaultMinutesToCheck { get; set; }
+        public double DefaultMinutesToCheck { get; set; }
     }
 
     public class SlackNotificationsSettings
